connectors/slack/requirements.txt
slack_bolt
requests
uvicorn
python-dotenv
slack